Preliminary implementation of docker support
This lets us load .tar files produced by docker save and run them,
the deployment manager need only send the tar file.

I think this is probably the lowest quality PR of the series, but
I'd like to get this done before the demo. Be harsh with your
criticism!

(On deck is docker support where we execute arbitrary commands!)